requirements.txt
aniso8601==9.0.1
attrs==24.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
Flask==2.2.5
Flask-Cors==5.0.0
flask-mongoengine==1.0.0
flask-restplus==0.13.0
flask-wtf==1.2.2
gunicorn==23.0.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mongoengine==0.29.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rpds-py==0.21.0
scipy==1.13.1
six==1.16.0
tzdata==2024.2
werkzeug==3.1.2
wtforms==3.2.1
zipp==3.21.0